feat: move AboutSection and HeaderSection components to specific folder
