+++ src/app/page.tsx
@@ -1,6 +1,6 @@
-import AboutSection from "@/components/about";
+import AboutSection from "@/components/about/about";
 import { Feature } from "@/components/feature/feature";
-import HeaderSection from "@/components/header";
+import HeaderSection from "@/components/header/header";
 
 export default function Home() {
   return (
